tweak test generator and add svg circle test
