+++ packages/app/integration-tests/tests/generate.js
@@ -1,0 +1,2 @@
+require('@babel/register');
+require('./generate-sandboxes');
