+++ packages/app/package.json
@@ -262,6 +262,7 @@
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
+    "@babel/register": "^7.13.16",
     "@divyenduz/ts-graphql-plugin": "^0.1.0",
     "@graphql-codegen/cli": "^1.15.4",
     "@graphql-codegen/fragment-matcher": "^1.15.4",
@@ -323,11 +324,11 @@
     "html-webpack-plugin": "^4.0.0-beta.1",
     "http-proxy-middleware": "^1.1.2",
     "imagemin-webpack-plugin": "^2.4.2",
+    "isomorphic-untar-gzip": "^1.0.5",
     "jest": "^24.9.0",
     "jest-image-snapshot": "^1.0.1",
     "jest-junit": "^3.4.1",
     "jest-styled-components": "7.0.2",
-    "isomorphic-untar-gzip": "^1.0.5",
     "mini-css-extract-plugin": "^0.8.0",
     "object-assign": "^4.1.1",
     "opn": "4.0.2",
